+++ src/lib-components/ButtonLink.vue
@@ -109,13 +109,28 @@ export default {
 </script>
 
 <template>
-  <SmartLink :to="to" :class="classes" :is-download="isDownload" :link-target="linkTarget">
+  <SmartLink
+    :to="to"
+    :class="classes"
+    :is-download="isDownload"
+    :link-target="linkTarget"
+  >
     <span class="label">{{ label }}</span>
 
-    <component :is="parsedIconName" class="arrow" aria-hidden="true" />
+    <component
+      :is="parsedIconName"
+      v-if="parsedIconName !== ''"
+      class="arrow"
+      aria-hidden="true"
+    />
     <div class="hover">
       <span class="label">{{ label }}</span>
-      <component :is="parsedIconName" class="arrow" aria-hidden="true" />
+      <component
+        :is="parsedIconName"
+        v-if="parsedIconName !== ''"
+        class="arrow"
+        aria-hidden="true"
+      />
     </div>
   </SmartLink>
 </template>