+++ src/stories/BlockForm.stories.js
@@ -1,3 +1,4 @@
+import { computed } from 'vue'
 import BlockForm from '@/lib-components/BlockForm'
 
 // Import mock api data
@@ -9,19 +10,21 @@ export default {
   component: BlockForm,
 }
 
+// The blockform data is tied to eventid, we cannot change eventid and keep the same form, so if you are updating form , also update the eventid and other details
 // Variations of stories below
 export function Default() {
   return {
     data() {
       return { BlockFormData }
     },
-    provide: {
-      // explicitly provide a computed property
-      eventId: '9383207',
-      blockFormData: BlockFormData.mock0,
-      libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
-      libcalWaitlist: '',
-      registrationType: 'both',
+    provide() {
+      return {
+        eventId: computed(() => '12333405'),
+        blockFormData: computed(() => BlockFormData.mock1),
+        libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
+        libcalWaitlist: computed(() => '1'),
+        registrationType: computed(() => 'in-person'),
+      }
     },
     components: { BlockForm },
     template: `
@@ -30,18 +33,19 @@ export function Default() {
   }
 }
 
-export function DefaultType2() {
+export function AppsTeamForm() {
   return {
     data() {
       return { BlockFormData }
     },
-    provide: {
-      // explicitly provide a computed property
-      eventId: '9383207',
-      blockFormData: BlockFormData.mock1,
-      libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
-      registrationType: 'both',
-      libcalWaitlist: null,
+    provide() {
+      return {
+        eventId: computed(() => '12333409'),
+        blockFormData: computed(() => BlockFormData.mock2),
+        libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
+        registrationType: computed(() => 'online'),
+        libcalWaitlist: computed(() => null),
+      }
     },
     components: { BlockForm },
     template: `
@@ -55,13 +59,14 @@ export function CLICCInstructionalUseForm() {
     data() {
       return { BlockFormData }
     },
-    provide: {
-      // explicitly provide a computed property
-      eventId: '9383207',
-      blockFormData: BlockFormData.mock2,
-      libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
-      registrationType: 'online',
-      libcalWaitlist: '0',
+    provide() {
+      return {
+        eventId: computed(() => '12333413'),
+        blockFormData: computed(() => BlockFormData.mock3),
+        libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
+        registrationType: computed(() => 'both'),
+        libcalWaitlist: computed(() => null),
+      }
     },
     components: { BlockForm },
     template: `
@@ -70,18 +75,19 @@ export function CLICCInstructionalUseForm() {
   }
 }
 
-export function TestFormTerryAllen() {
+export function LibrarySpecialCollections() {
   return {
     data() {
       return { BlockFormData }
     },
-    provide: {
-      // explicitly provide a computed property
-      eventId: '9383207',
-      blockFormData: BlockFormData.mock3,
-      libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
-      registrationType: 'in-person',
-      libcalWaitlist: '1',
+    provide() {
+      return {
+        eventId: computed(() => '12333417'),
+        blockFormData: computed(() => BlockFormData.mock4),
+        libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
+        registrationType: computed(() => 'online'),
+        libcalWaitlist: computed(() => null),
+      }
     },
     components: { BlockForm },
     template: `
@@ -90,18 +96,40 @@ export function TestFormTerryAllen() {
   }
 }
 
-export function TestFormFTVA() {
+export function FTVATest() {
   return {
     data() {
       return { BlockFormData }
     },
-    provide: {
-      // explicitly provide a computed property
-      eventId: '9383207',
-      blockFormData: BlockFormData.mock4,
-      libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
-      registrationType: 'in-person',
-      libcalWaitlist: '',
+    provide() {
+      return {
+        eventId: computed(() => '12333419'),
+        blockFormData: computed(() => BlockFormData.mock5),
+        libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
+        registrationType: computed(() => 'in-person'),
+        libcalWaitlist: computed(() => '1'),
+      }
+    },
+    components: { BlockForm },
+    template: `
+        <block-form />
+    `,
+  }
+}
+
+export function TerryAllen() {
+  return {
+    data() {
+      return { BlockFormData }
+    },
+    provide() {
+      return {
+        eventId: computed(() => '12333423'),
+        blockFormData: computed(() => BlockFormData.mock6),
+        libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
+        registrationType: computed(() => 'online'),
+        libcalWaitlist: computed(() => null),
+      }
     },
     components: { BlockForm },
     template: `
@@ -115,13 +143,14 @@ export function TextBallot() {
     data() {
       return { BlockFormData }
     },
-    provide: {
-      // explicitly provide a computed property
-      eventId: '9383207',
-      blockFormData: BlockFormData.mock5,
-      libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
-      registrationType: 'in-person',
-      libcalWaitlist: null,
+    provide() {
+      return {
+        eventId: computed(() => '12333425'),
+        blockFormData: computed(() => BlockFormData.mock7),
+        libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
+        registrationType: computed(() => 'in-person'),
+        libcalWaitlist: computed(() => null),
+      }
     },
     components: { BlockForm },
     template: `
@@ -130,18 +159,19 @@ export function TextBallot() {
   }
 }
 
-export function UCLALibraryEventRegistration() {
+export function UCLALibrary() {
   return {
     data() {
       return { BlockFormData }
     },
-    provide: {
-      // explicitly provide a computed property
-      eventId: '9383207',
-      blockFormData: BlockFormData.mock6,
-      libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
-      registrationType: 'in-person',
-      libcalWaitlist: '1',
+    provide() {
+      return {
+        eventId: computed(() => '12333491'),
+        blockFormData: computed(() => BlockFormData.mock8),
+        libcalEndpoint: 'https://proxy.calendar.library.ucla.edu/',
+        registrationType: computed(() => 'both'),
+        libcalWaitlist: computed(() => null),
+      }
     },
     components: { BlockForm },
     template: `